move plotting.img_plotting under plotting.image
