+++ nilearn/plotting/image/__init__.py
@@ -1,0 +1,27 @@
+from nilearn.plotting.image.img_plotting import (
+    plot_anat,
+    plot_carpet,
+    plot_connectome,
+    plot_epi,
+    plot_glass_brain,
+    plot_img,
+    plot_markers,
+    plot_prob_atlas,
+    plot_roi,
+    plot_stat_map,
+    show,
+)
+
+__all__ = [
+    "plot_anat",
+    "plot_carpet",
+    "plot_connectome",
+    "plot_epi",
+    "plot_glass_brain",
+    "plot_img",
+    "plot_markers",
+    "plot_prob_atlas",
+    "plot_roi",
+    "plot_stat_map",
+    "show",
+]
